Criados Posts e Gets

+++ src/main/java/com/bradesco/remessasPJ/repository/CarteiraRepository.java
@@ -3,5 +3,5 @@ package com.bradesco.remessasPJ.repository;
 import com.bradesco.remessasPJ.entidade.Carteira;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarteiraRepository extends JpaRepository<Carteira, Long> {
+public interface CarteiraRepository extends JpaRepository<Carteira, Integer> {
 }
