+++ src/main/java/com/bradesco/remessasPJ/repository/ContratoRepository.java
@@ -3,5 +3,5 @@ package com.bradesco.remessasPJ.repository;
 import com.bradesco.remessasPJ.entidade.Contrato;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContratoRepository extends JpaRepository<Contrato, Long> {
+public interface ContratoRepository extends JpaRepository<Contrato, Integer> {
 }
