+++ src/main/java/com/bradesco/remessasPJ/repository/RemessaRepository.java
@@ -3,5 +3,5 @@ package com.bradesco.remessasPJ.repository;
 import com.bradesco.remessasPJ.entidade.Remessa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RemessaRepository extends JpaRepository<Remessa, Long> {
+public interface RemessaRepository extends JpaRepository<Remessa, Integer> {
 }
