+++ src/main/java/com/bradesco/remessasPJ/repository/StatusRepository.java
@@ -3,5 +3,5 @@ package com.bradesco.remessasPJ.repository;
 import com.bradesco.remessasPJ.entidade.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Long> {
+public interface StatusRepository extends JpaRepository<Status, Integer> {
 }
